+++ modules/core/dependency/python-ihm/ihm/format.py
@@ -278,12 +278,25 @@ class CifReader(_Reader):
               not present in the file or is the mmCIF omitted value (.).
               (mmCIF keywords are case insensitive, so this class always treats
               them as lowercase regardless of the file contents.)
+       :param unknown_category_handler: A callable (or `None`) that is called
+              for each category in the file that isn't handled; it is given
+              two arguments: the name of the category, and the line in the
+              file at which the category was encountered (if known, otherwise
+              None).
+       :param unknown_keyword_handler: A callable (or `None`) that is called
+              for each keyword in the file that isn't handled (within a category
+              that is handled); it is given three arguments: the names of the
+              category and keyword, and the line in the file at which the
+              keyword was encountered (if known, otherwise None).
     """
-    def __init__(self, fh, category_handler):
+    def __init__(self, fh, category_handler, unknown_category_handler=None,
+                 unknown_keyword_handler=None):
         if _format is not None:
             c_file = _format.ihm_file_new_from_python(fh)
             self._c_format = _format.ihm_reader_new(c_file)
         self.category_handler = category_handler
+        self.unknown_category_handler = unknown_category_handler
+        self.unknown_keyword_handler = unknown_keyword_handler
         self._category_data = {}
         self.fh = fh
         self._tokens = []
@@ -405,44 +418,55 @@ class CifReader(_Reader):
     def _read_value(self, vartoken):
         """Read a line that sets a single value, e.g. "_entry.id   1YTI"""
         # Only read the value if we're interested in this category and key
-        if vartoken.category in self.category_handler \
-          and vartoken.keyword \
-          in self.category_handler[vartoken.category]._keys:
-            valtoken = self._get_token()
-            if isinstance(valtoken, _ValueToken):
-                ch = self.category_handler[vartoken.category]
-                if vartoken.category not in self._category_data:
-                    self._category_data[vartoken.category] = {}
-                if valtoken.txt == '.':
-                    val = ch.omitted
-                elif valtoken.txt == '?':
-                    val = ch.unknown
+        if vartoken.category in self.category_handler:
+            if vartoken.keyword \
+               in self.category_handler[vartoken.category]._keys:
+                valtoken = self._get_token()
+                if isinstance(valtoken, _ValueToken):
+                    ch = self.category_handler[vartoken.category]
+                    if vartoken.category not in self._category_data:
+                        self._category_data[vartoken.category] = {}
+                    if valtoken.txt == '.':
+                     val = ch.omitted
+                    elif valtoken.txt == '?':
+                        val = ch.unknown
+                    else:
+                        val = valtoken.txt
+                    self._category_data[vartoken.category][vartoken.keyword] \
+                                           = val
                 else:
-                    val = valtoken.txt
-                self._category_data[vartoken.category][vartoken.keyword] = val
-            else:
-                raise CifParserError("No valid value found for %s.%s on line %d"
-                              % (vartoken.category, vartoken.keyword,
-                                 self._linenum))
+                    raise CifParserError(
+                               "No valid value found for %s.%s on line %d"
+                               % (vartoken.category, vartoken.keyword,
+                                  self._linenum))
+            elif self.unknown_keyword_handler is not None:
+                self.unknown_keyword_handler(vartoken.category,
+                                             vartoken.keyword, self._linenum)
+        elif self.unknown_category_handler is not None:
+            self.unknown_category_handler(vartoken.category, self._linenum)
 
     def _read_loop_keywords(self):
         """Read the set of keywords for a loop_ construct"""
         category = None
         keywords = []
+        first_line = None
+        keyword_lines = []
         while True:
             token = self._get_token()
             if isinstance(token, _VariableToken):
                 if category is None:
                     category = token.category
+                    first_line = self._linenum
                 elif category != token.category:
                     raise CifParserError("mmCIF files cannot contain multiple "
                                 "categories within a single loop at line %d"
                                 % self._linenum)
                 keywords.append(token.keyword)
+                keyword_lines.append(self._linenum)
             elif isinstance(token, _ValueToken):
                 # OK, end of keywords; proceed on to values
                 self._unget_token()
-                return category, keywords
+                return category, keywords, keyword_lines, first_line
             else:
                 raise CifParserError("Was expecting a keyword or value for "
                                      "loop at line %d" % self._linenum)
@@ -473,7 +497,8 @@ class CifReader(_Reader):
 
     def _read_loop(self):
         """Handle a loop_ construct"""
-        category, keywords = self._read_loop_keywords()
+        (category, keywords,
+                keyword_lines, first_line) = self._read_loop_keywords()
         # Skip data if we don't have a handler for it
         if category in self.category_handler:
             ch = self.category_handler[category]
@@ -481,7 +506,13 @@ class CifReader(_Reader):
             for i, k in enumerate(ch._keys):
                 wanted_key_index[k] = i
             indices = [wanted_key_index.get(k, -1) for k in keywords]
+            if self.unknown_keyword_handler is not None:
+                for k, i, line in zip(keywords, indices, keyword_lines):
+                    if i == -1:
+                        self.unknown_keyword_handler(category, k, line)
             self._read_loop_data(ch, len(ch._keys), indices)
+        elif self.unknown_category_handler is not None:
+            self.unknown_category_handler(category, first_line)
 
     def read_file(self):
         """Read the file and extract data.
@@ -542,6 +573,12 @@ class CifReader(_Reader):
             func = getattr(handler, '_add_c_handler', None) \
                         or _format.add_category_handler
             func(self._c_format, category, handler._keys, handler)
+        if self.unknown_category_handler is not None:
+            _format.add_unknown_category_handler(self._c_format,
+                                                 self.unknown_category_handler)
+        if self.unknown_keyword_handler is not None:
+            _format.add_unknown_keyword_handler(self._c_format,
+                                                self.unknown_keyword_handler)
         try:
             eof, more_data = _format.ihm_read_file(self._c_format)
         except _format.FileFormatError as exc: